AUT-5466: Change passkeyCounter to long
Matches the type on the assertion result

+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/services/webauthn/PasskeyAssertionService.java
@@ -119,7 +119,7 @@ public class PasskeyAssertionService {
         var passkeyDetail =
                 PasskeyDetail.verificationFailed(
                         userVerification,
-                        (int) assertionResult.getSignatureCount(), // TODO change to long
+                        assertionResult.getSignatureCount(),
                         assertionResult.isBackedUp(),
                         passkeyCredentialDeviceType,
                         "UserVerificationError");
